+++ receiver/kafkametricsreceiver/kafkametrics_e2e_test.go
@@ -56,8 +56,8 @@ func TestIntegrationSingleNode(t *testing.T) {
 	f := NewFactory()
 	cfg := f.CreateDefaultConfig().(*Config)
 	cfg.Scrapers = []string{
-		"consumers",
 		"brokers",
+		"consumers",
 		"topics",
 	}
 	cfg.Brokers = []string{kafkaAddress}
@@ -67,16 +67,12 @@ func TestIntegrationSingleNode(t *testing.T) {
 
 	var receiver component.MetricsReceiver
 	var err error
-	t.Logf("waiting to connect to kafka...")
-	require.Eventuallyf(t,
-		func() bool {
-			receiver, err = f.CreateMetricsReceiver(context.Background(), params, cfg, consumer)
-			return err == nil
-		}, 30*time.Second, 5*time.Second,
-		fmt.Sprintf("failed to create metrics receiver. %v", err),
-	)
-	t.Logf("connected to kafka")
-	require.NoError(t, receiver.Start(context.Background(), &testHost{t: t}))
+	receiver, err = f.CreateMetricsReceiver(context.Background(), params, cfg, consumer)
+	require.NoError(t, err, "failed to create receiver")
+	require.Eventuallyf(t, func() bool {
+		err = receiver.Start(context.Background(), &testHost{t: t})
+		return err == nil
+	}, 30*time.Second, 5*time.Second, fmt.Sprintf("failed to start metrics receiver. %v", err))
 	t.Logf("waiting for metrics...")
 	require.Eventuallyf(t,
 		func() bool {